Use a custom supervisor for ranch_conns_sup
This change was designed so that we don't have this supervisor
and ranch_listener performing the same job, namely monitoring
connection processes, the first through links and the second
through monitors.

This change also makes possible various optimizations:

 *  Acceptors don't need to know about options, maximum number
    of connections, or anything else. They can just accept,
    pass the socket to the supervisor, and when the supervisor
    replies continue accepting connections.

 *  The supervisor holds most of the information that will be
    passed to created processes. This reduces copying.

 *  The supervisor temporarily takes ownership of the socket,
    then creates the connection process and gives it ownership,
    streamlining the creation.

 *  The supervisor can hold acceptors in their receive loop if
    max_connections is reached. When this number gets below the
    limit it can then send a message to a sleeping acceptor to
    make it resume its operations.

 *  Because we know that all connection process creations are made
    from the local Erlang node, we can greatly reduce the number
    operations to be made when calling the supervisor.

 *  Because all acceptors die if this supervisor dies, we can
    remove even more operations from the calling code. We do not
    need to monitor or wait for a timeout. This reduces the call
    code to two statements: send and receive. (Thanks James Fish
    for helping with that.)

 *  The supervisor only needs to keep track of a list of pids.
    There is no children specification to be maintained, we do
    not need to handle restart strategy (no process can be
    restarted because the socket dies with it). We are using
    the process dictionary for storing the pids as it proved
    to be the simplest and fastest solution.

 *  The supervisor maintains a count of current connections,
    but also of processes (including the ones that removed
    themselves from the pool), making any query of these values
    very fast.

The supervisor should still be compatible with OTP principles.
It responds to calls from the supervisor module as expected,
although some of them are disabled and an error will be returned,
for example supervisor:start_child/2. It is also started with
proc_lib and handles system messages. sys:get_status/1 can thus
be used as expected.

We can see a great increase in the number of requests/s, a great
improvement in the latency of requests, and we can simply accept
requests faster than before. It will probably have a bigger increase
under virtualized environments, although that's only a guess.

As a result of this, we don't write much anymore in the ranch_server
ets table, so the write_concurrency option was removed. Tests were
also slightly improved to prevent race conditions.