Work around broken elliptic-curve cipher suites
Unfortunately the implementation of elliptic-curve ciphers that has
been introduced in R16B01 is incomplete.  Depending on the particular
client, this can cause the TLS handshake to break during key agreement.
As it turns out that most popular browsers (e.g. Firefox, Chromium,
and Safari) are affected by this bug, we provide this workaround.

This workaround makes sure that only cipher suite implementations that
are not known to be broken are supported by default.