Better distinguish between Ranch and socket options
A map should now be used when specifying transport options
that contain more than just socket options. It is still
possible to pass a list of socket options directly as a
convenience.

The ack_timeout is renamed to handshake_timeout when
specified as a map. This corresponds to the new function
ranch:handshake/1,2 that will be favored in Ranch 2.0.

Specifying Ranch-specific options via the proplist will
no longer be possible starting from Ranch 2.0.