Clean listener options after normal shutdown
In addition to cleaning when ranch:stop_listener/1 is called, we
also need to clean when we detect the supervisor is going away
for normal reasons, because the supervisor might be in another
application's supervision tree.

Note that there might be a short delay in this case before the
cleanup is done, due to using monitors for detection.